+++ src/main/resources/data/computercraft/lua/rom/apis/rednet.lua
@@ -193,7 +193,7 @@ function send(recipient, message, protocol)
     local sent = false
     if recipient == os.getComputerID() then
         -- Loopback to ourselves
-        os.queueEvent("rednet_message", os.getComputerID(), message_wrapper, protocol)
+        os.queueEvent("rednet_message", os.getComputerID(), message, protocol)
         sent = true
     else
         -- Send on all open modems, to the target and to repeaters
@@ -217,9 +217,10 @@ end
 channel. The message will be received by every device listening to rednet.
 
 @param message The message to send. This should not contain coroutines or
-functions, as they will be converted to @{nil}.  @tparam[opt] string protocol
-The "protocol" to send this message under. When using @{rednet.receive} one can
-filter to only receive messages sent under a particular protocol.
+functions, as they will be converted to @{nil}.
+@tparam[opt] string protocol The "protocol" to send this message under. When
+using @{rednet.receive} one can filter to only receive messages sent under a
+particular protocol.
 @see rednet.receive
 @changed 1.6 Added protocol parameter.
 @usage Broadcast the words "Hello, world!" to every computer using rednet.
@@ -318,7 +319,7 @@ different, or if they only join a given network after "registering" themselves
 before doing so (eg while offline or part of a different network).
 
 @tparam string protocol The protocol this computer provides.
-@tparam string hostname The name this protocol exposes for the given protocol.
+@tparam string hostname The name this computer exposes for the given protocol.
 @throws If trying to register a hostname which is reserved, or currently in use.
 @see rednet.unhost
 @see rednet.lookup