+++ src/main/resources/data/computercraft/lua/rom/programs/import.lua
@@ -18,7 +18,5 @@ if #files == 0 then
   return
 end
 
-package.path = package.path .. "/rom/modules/internal/?.lua"
-
-local ok, err = require("cc.import")(files)
+local ok, err = require("cc.internal.import")(files)
 if not ok and err then printError(err) end