+++ src/test/java/dan200/computercraft/core/LuaCoverage.java
@@ -6,9 +6,12 @@
 package dan200.computercraft.core;
 
 import com.google.common.base.Strings;
-import dan200.computercraft.ComputerCraft;
+import it.unimi.dsi.fastutil.ints.Int2IntMap;
+import it.unimi.dsi.fastutil.ints.Int2IntMaps;
 import it.unimi.dsi.fastutil.ints.IntOpenHashSet;
 import it.unimi.dsi.fastutil.ints.IntSet;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.squiddev.cobalt.Prototype;
 import org.squiddev.cobalt.compiler.CompileException;
 import org.squiddev.cobalt.compiler.LuaC;
@@ -16,30 +19,27 @@ import org.squiddev.cobalt.compiler.LuaC;
 import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.*;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
+import java.util.ArrayDeque;
+import java.util.Collections;
+import java.util.Map;
+import java.util.Queue;
 
 class LuaCoverage
 {
+    private static final Logger LOG = LoggerFactory.getLogger( LuaCoverage.class );
     private static final Path ROOT = new File( "src/main/resources/data/computercraft/lua" ).toPath();
-    private static final Path BIOS = ROOT.resolve( "bios.lua" );
-    private static final Path APIS = ROOT.resolve( "rom/apis" );
-    private static final Path SHELL = ROOT.resolve( "rom/programs/shell.lua" );
-    private static final Path MULTISHELL = ROOT.resolve( "rom/programs/advanced/multishell.lua" );
-    private static final Path TREASURE = ROOT.resolve( "treasure" );
 
-    private final Map<String, Map<Double, Double>> coverage;
+    private final Map<String, Int2IntMap> coverage;
     private final String blank;
     private final String zero;
     private final String countFormat;
 
-    LuaCoverage( Map<String, Map<Double, Double>> coverage )
+    LuaCoverage( Map<String, Int2IntMap> coverage )
     {
         this.coverage = coverage;
 
-        int max = (int) coverage.values().stream()
-            .flatMapToDouble( x -> x.values().stream().mapToDouble( y -> y ) )
+        int max = coverage.values().stream()
+            .flatMapToInt( x -> x.values().stream().mapToInt( y -> y ) )
             .max().orElse( 0 );
         int maxLen = Math.max( 1, (int) Math.ceil( Math.log10( max ) ) );
         blank = Strings.repeat( " ", maxLen + 1 );
@@ -49,25 +49,22 @@ class LuaCoverage
 
     void write( Writer out ) throws IOException
     {
-        Files.find( ROOT, Integer.MAX_VALUE, ( path, attr ) -> attr.isRegularFile() && !path.startsWith( TREASURE ) ).forEach( path -> {
+        Files.find( ROOT, Integer.MAX_VALUE, ( path, attr ) -> attr.isRegularFile() ).forEach( path -> {
             Path relative = ROOT.relativize( path );
             String full = relative.toString().replace( '\\', '/' );
             if( !full.endsWith( ".lua" ) ) return;
 
-            Map<Double, Double> files = Stream.of(
-                coverage.remove( "/" + full ),
-                path.equals( BIOS ) ? coverage.remove( "bios.lua" ) : null,
-                path.equals( SHELL ) ? coverage.remove( "shell.lua" ) : null,
-                path.equals( MULTISHELL ) ? coverage.remove( "multishell.lua" ) : null,
-                path.startsWith( APIS ) ? coverage.remove( path.getFileName().toString() ) : null
-            )
-                .filter( Objects::nonNull )
-                .flatMap( x -> x.entrySet().stream() )
-                .collect( Collectors.toMap( Map.Entry::getKey, Map.Entry::getValue, Double::sum ) );
+            Int2IntMap possiblePaths = coverage.remove( "/" + full );
+            if( possiblePaths == null ) possiblePaths = coverage.remove( full );
+            if( possiblePaths == null )
+            {
+                possiblePaths = Int2IntMaps.EMPTY_MAP;
+                LOG.warn( "{} has no coverage data", full );
+            }
 
             try
             {
-                writeCoverageFor( out, path, files );
+                writeCoverageFor( out, path, possiblePaths );
             }
             catch( IOException e )
             {
@@ -78,15 +75,15 @@ class LuaCoverage
         for( String filename : coverage.keySet() )
         {
             if( filename.startsWith( "/test-rom/" ) ) continue;
-            ComputerCraft.log.warn( "Unknown file {}", filename );
+            LOG.warn( "Unknown file {}", filename );
         }
     }
 
-    private void writeCoverageFor( Writer out, Path fullName, Map<Double, Double> visitedLines ) throws IOException
+    private void writeCoverageFor( Writer out, Path fullName, Int2IntMap visitedLines ) throws IOException
     {
         if( !Files.exists( fullName ) )
         {
-            ComputerCraft.log.error( "Cannot locate file {}", fullName );
+            LOG.error( "Cannot locate file {}", fullName );
             return;
         }
 
@@ -104,10 +101,10 @@ class LuaCoverage
             while( (line = reader.readLine()) != null )
             {
                 lineNo++;
-                Double count = visitedLines.get( (double) lineNo );
-                if( count != null )
+                int count = visitedLines.getOrDefault( lineNo, -1 );
+                if( count >= 0 )
                 {
-                    out.write( String.format( countFormat, count.intValue() ) );
+                    out.write( String.format( countFormat, count ) );
                 }
                 else if( activeLines.contains( lineNo ) )
                 {
@@ -128,28 +125,30 @@ class LuaCoverage
     private static IntSet getActiveLines( File file ) throws IOException
     {
         IntSet activeLines = new IntOpenHashSet();
-        try( InputStream stream = new FileInputStream( file ) )
+        Queue<Prototype> queue = new ArrayDeque<>();
+
+        try( InputStream stream = Files.newInputStream( file.toPath() ) )
         {
             Prototype proto = LuaC.compile( stream, "@" + file.getPath() );
-            Queue<Prototype> queue = new ArrayDeque<>();
             queue.add( proto );
-
-            while( (proto = queue.poll()) != null )
-            {
-                int[] lines = proto.lineinfo;
-                if( lines != null )
-                {
-                    for( int line : lines )
-                    {
-                        activeLines.add( line );
-                    }
-                }
-                if( proto.p != null ) Collections.addAll( queue, proto.p );
-            }
         }
         catch( CompileException e )
         {
             throw new IllegalStateException( "Cannot compile", e );
+        }
+
+        Prototype proto;
+        while( (proto = queue.poll()) != null )
+        {
+            int[] lines = proto.lineInfo;
+            if( lines != null )
+            {
+                for( int line : lines )
+                {
+                    activeLines.add( line );
+                }
+            }
+            if( proto.children != null ) Collections.addAll( queue, proto.children );
         }
 
         return activeLines;