+++ src/test/resources/test-rom/spec/programs/turtle/unequip_spec.lua
@@ -4,7 +4,7 @@ describe("The turtle unequip program", function()
     it("errors when not a turtle", function()
         stub(_G, "turtle", nil)
 
-        expect(capture(stub, "/rom/programs/turtle/unequip.lua"))
+        expect(capture("/rom/programs/turtle/unequip.lua"))
             :matches { ok = true, output = "", error = "Requires a Turtle\n" }
     end)
 
@@ -12,7 +12,7 @@ describe("The turtle unequip program", function()
     it("displays its usage when given no arguments", function()
         stub(_G, "turtle", {})
 
-        expect(capture(stub, "/rom/programs/turtle/unequip.lua"))
+        expect(capture("/rom/programs/turtle/unequip.lua"))
             :matches { ok = true, output = "Usage: /rom/programs/turtle/unequip.lua <side>\n", error = "" }
     end)
 
@@ -24,9 +24,9 @@ describe("The turtle unequip program", function()
             equipLeft = function() return true end,
         })
 
-        expect(capture(stub, "/rom/programs/turtle/unequip.lua left"))
+        expect(capture("/rom/programs/turtle/unequip.lua left"))
             :matches { ok = true, output = "Nothing to unequip\n", error = "" }
-        expect(capture(stub, "/rom/programs/turtle/unequip.lua right"))
+        expect(capture("/rom/programs/turtle/unequip.lua right"))
             :matches { ok = true, output = "Nothing to unequip\n", error = "" }
     end)
 
@@ -45,10 +45,10 @@ describe("The turtle unequip program", function()
             end,
         })
 
-        expect(capture(stub, "/rom/programs/turtle/unequip.lua left"))
+        expect(capture("/rom/programs/turtle/unequip.lua left"))
             :matches { ok = true, output = "Item unequipped\n", error = "" }
         item_count = 0
-        expect(capture(stub, "/rom/programs/turtle/unequip.lua right"))
+        expect(capture("/rom/programs/turtle/unequip.lua right"))
             :matches { ok = true, output = "Item unequipped\n", error = "" }
     end)
 
@@ -60,9 +60,9 @@ describe("The turtle unequip program", function()
             equipLeft = function() return true end,
         })
 
-        expect(capture(stub, "/rom/programs/turtle/unequip.lua left"))
+        expect(capture("/rom/programs/turtle/unequip.lua left"))
             :matches { ok = true, output = "No space to unequip item\n", error = "" }
-        expect(capture(stub, "/rom/programs/turtle/unequip.lua right"))
+        expect(capture("/rom/programs/turtle/unequip.lua right"))
             :matches { ok = true, output = "No space to unequip item\n", error = "" }
     end)